requirements.txt
Django==1.3.1
Markdown==2.1.1
django-pdb==0.2.3
django-tagging==0.3.1
elementtree==1.2.7-20070827-preview
wsgiref==0.1.2
